미체결 조회, 삭제 로직 구현

+++ src/main/java/bit/bitgroundspring/service/OrderExecutionService.java
@@ -81,7 +81,7 @@ public class OrderExecutionService {
             
             Order order = orderOpt.get();
             order.setStatus(Status.COMPLETED);
-            order.setTradePrice((float) executionPrice);
+            order.setTradePrice((double) executionPrice);
             order.setUpdatedAt(LocalDateTime.now());
             
             orderRepository.save(order);
@@ -92,7 +92,7 @@ public class OrderExecutionService {
             String symbol = coinRepository.findById(order.getCoin().getId())
                     .map(Coin::getSymbol)
                     .orElse("Unknown");
-            float amount = order.getAmount();
+            Double amount = order.getAmount();
             float tradePrice = order.getTradePrice().intValue();
             Map<String, Object> data = Map.of(
                     "orderType", orderType.name(),
@@ -115,4 +115,5 @@ public class OrderExecutionService {
             log.error("Failed to execute order: {}", orderId, e);
         }
     }
+
 }
